Add new command "warn [DEVICE] [WARNINGMODE] [STROBE] [DURATION]" to start warning a specfied IAS warning device

+++ zigbee-api/src/main/java/org/bubblecloud/zigbee/api/cluster/impl/api/security_safety/ias_wd/StartWarningPayload.java
@@ -32,5 +32,5 @@ public interface StartWarningPayload {
 
     public short getStrobe();
 
-    public int getWarningDuration();
+    public short getWarningDuration();
 }
